warehouse/backend/requirements.txt
torch>=1.9.0
numpy>=1.21.0
matplotlib>=3.3.0
scipy>=1.7.0
pillow>=8.0.0
pandas>=1.3.0
tensorflow>=2.8.0
